Interview list wrapper + alt text and CSS (#10)
Fix #9

Really wasn't anything else to add to the interview list wrapper, just to include the `interview_list.yml` and `mlh_interview_framework.yml`.

I also added alt text to the logo (we can always change it later, but it's a good placeholder to say where the link will lead to), and corrected some CSS issues, that just happened to be working, but we really being ignored. `#` makes a comment in python, but in CSS you have to wrap everything in the comment with `/*` and `*/`. 🤷🏻