CSS changes (#12)
This was more low hanging fruit that I figured I could help set up: some quick colors and fonts.

I based everything off https://michiganlegalhelp.org/guide-to-legal-help, happy to edit or change anything later. I used the [ALDashboard bootstrap theme generator](https://suffolklitlab.org/docassemble-AssemblyLine-documentation/docs/customization/overview/#creating-a-custom-css-theme-with-bootstrapbuild), keeping most of the DA defaults except the primary color (which is now that MLH dark blue), the warning yellow (now a similar yellow from the MLH site) and the danger red (which is the red used in the above linked online form).

Since the logo was the same color as that MLH dark blue, I added a little white rounded background, kinda like an app button. Happy to change that too.

I also used the same fonts, with the same font size and line height. I did shrink the section label font size, so more of them can fit on a single line for legibility. These will likely have to be changed, but I figured getting something in that's easily editable is better than trying to figure out where to change it later.

I included the original scss file, so when we want to make changes, we can edit that file, and upload it to the Dashboard bootstrap theme tool. In that SCSS, I set the variables outside of the theme, which aligns many different colors across bootstrap, so we don't have some components with link blue and others with our blue.

Also included a workaround for a docassemble bug when including the bootstrap theme. Can't test in playground as easily, but that's okay for now.