Turn navigation buttons into links
From an accessibility audit of some of Vermont's docassemble forms:

> Some `button` elements should really be links since their function is navigation; for example, “Go to
  legal aid website” ["Exit" in our case] if the user can’t continue the interview. We recommend changing all
  elements having a navigation purpose to links. 3-Moderate

Some more details about using links for navigation and buttons for form logic
are in https://ashleemboyer.com/blog/should-i-use-a-button-or-a-link.

We use "Exit" buttons instead of the recommended links as well, and this PR
changes the exit button in the framework (other similar changes are coming for
SimpMotion and Motion Regarding Child Support).