Remove unused exit link from default screen parts
[Exit link](https://docassemble.org/docs/initial.html#exit%20link) defines
the behavior of the top right exit button when `show login` is False, something
that we haven't done, and I don't think we will do for our interviews.

Given we had already incorrectly set it to a URL, and I got confused and
made incorrect comments on three different issues because it's a strangely
described feature, IMO we should just remove it. Not urgent though.